requirements.txt
pydub
telebot
openai
openpyxl
pandas
python-dateutil
pymongo
python-dotenv
dnspython
google-cloud-texttospeech
